+++ Suspension/src/Suspension.Application/Controllers/ReactionController.cs
@@ -2,7 +2,9 @@
 using Microsoft.AspNetCore.Mvc;
 using MudRunner.Suspension.Application.Extensions;
 using MudRunner.Suspension.Core.Operations.CalculateReactions;
+using MudRunner.Suspension.Core.Operations.CalculateStearingKnuckleReactions;
 using MudRunner.Suspension.DataContracts.CalculateReactions;
+using MudRunner.Suspension.DataContracts.CalculateSteeringKnuckleReactions;
 using System.Threading.Tasks;
 
 namespace MudRunner.Suspension.Application.Controllers